Pull out flask app creation into it's own file and module
I'm doing this, because for testing I'll need a handle on
the app
