+++ .gitignore
@@ -1,1 +1,2 @@
 books.db
+__pycache__/
